Safari-609.1.11.1/LayoutTests/http/tests/appcache/resource-redirect-2-expected.txt

CONSOLE MESSAGE: line 1: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Application Cache update failed, because http://127.0.0.1:8000/appcache/resources/resource-redirect-2.php was redirected.
Test that a redirect makes resource caching fail.

Should say SUCCESS:

SUCCESS
